package/meson: bump to version 0.47.2
Note, that there is already a patch [1] to bump the version to 0.48.1.
However, the bump to version 0.48.x has some unresolved issues.

In the meantime, until the issue with 0.48 is resolved, bump to latest
version of the 0.47 branch.

[1] http://patchwork.ozlabs.org/patch/986260/

Signed-off-by: Jörg Krause <joerg.krause@embedded.rocks>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
(cherry picked from commit 55b5dfc1bf99999592ad074ef18f59ce92c69a76)
Signed-off-by: Cody Xie <cody.xie@rock-chips.com>
Change-Id: I98c0a06097980f553bfae8351a23be891c7f7369