Makefile: unexport 'PLATFORM' and 'OS' environment variables
Some package builds may fail when environment variables are present with the
same names as make variables in a package. This is a bigger problem for
environment variables with generic names, like 'PLATFORM' and 'OS'.

'PLATFORM' is for example a problem for host-acl.

Signed-off-by: Thomas De Schampheleire <thomas.de_schampheleire@nokia.com>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>
(cherry picked from commit d3e535a839a893d4721240278fada4652544b60c)
Signed-off-by: Jeffy Chen <jeffy.chen@rock-chips.com>

Change-Id: I7ddbae8a94ac407a18108a56087aebd082d0ad15