package/opencv: drop gstreamer 0.10.x support
With the upcoming removal of gstreamer 0.10, the logic for building opencv
with support for it must go as well.

As there is now a single option for gstreamer (1.x) support, convert the
gstreamer support choice to a normal option for simplicity.

Signed-off-by: Peter Korsgaard <peter@korsgaard.com>
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>
(cherry picked from commit 0220ad8b138da666fd546d749674dcbde1ecb12a)

Change-Id: If54f92bc61e598dfac1ea8810240bcd7aa866cd6
Signed-off-by: Jeffy Chen <jeffy.chen@rock-chips.com>