package/opencv3: drop gstreamer 0.10.x support
With the upcoming removal of gstreamer 0.10, the logic for building opencv3
with support for it must go as well.

As there is now a single option for gstreamer (1.x) support, convert the
gstreamer support choice to a normal option for simplicity.

Signed-off-by: Peter Korsgaard <peter@korsgaard.com>
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>
(cherry picked from commit ceb23e88c9d9db302c339778be2be5b785a6c984)

Change-Id: Ic3e82007b59a5f459ffd4892cc8fdfb652b74ef9
Signed-off-by: Jeffy Chen <jeffy.chen@rock-chips.com>