package/freerdp: drop gstreamer 0.10.x support
With the upcoming removal of gstreamer 0.10, the logic for building freerdp
with support for it must go as well.

As there is now a single option for gstreamer (1.x) support, convert the
gstreamer support choice to a normal option for simplicity.

Signed-off-by: Peter Korsgaard <peter@korsgaard.com>
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>
(cherry picked from commit 545ffdc1856d56eba2c4974559abdba918ba7f08)

Change-Id: I0dab4320e881e165f672ede76fd0bee3fb0e4ee3
Signed-off-by: Jeffy Chen <jeffy.chen@rock-chips.com>